requirements.txt
Flask==3.1.0
Flask_Cors==5.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
Requests==2.32.3
